Rename testutils to harness
This allows us to bring in a dependency named `testutils`.
