+++ swap/tests/happy_path.rs
@@ -1,14 +1,14 @@
-pub mod testutils;
+pub mod harness;
 
+use harness::SlowCancelConfig;
 use swap::protocol::{alice, bob};
-use testutils::SlowCancelConfig;
 use tokio::join;
 
 /// Run the following tests with RUST_MIN_STACK=10000000
 
 #[tokio::test]
 async fn happy_path() {
-    testutils::setup_test(SlowCancelConfig, |mut ctx| async move {
+    harness::setup_test(SlowCancelConfig, |mut ctx| async move {
         let (bob_swap, _) = ctx.bob_swap().await;
         let bob_swap = tokio::spawn(bob::run(bob_swap));
