+++ swap/tests/happy_path_restart_bob_before_xmr_locked.rs
@@ -1,13 +1,13 @@
-pub mod testutils;
+pub mod harness;
 
+use harness::bob_run_until::is_xmr_locked;
+use harness::SlowCancelConfig;
 use swap::protocol::bob::BobState;
 use swap::protocol::{alice, bob};
-use testutils::bob_run_until::is_xmr_locked;
-use testutils::SlowCancelConfig;
 
 #[tokio::test]
 async fn given_bob_restarts_after_xmr_is_locked_resume_swap() {
-    testutils::setup_test(SlowCancelConfig, |mut ctx| async move {
+    harness::setup_test(SlowCancelConfig, |mut ctx| async move {
         let (bob_swap, bob_join_handle) = ctx.bob_swap().await;
         let bob_swap = tokio::spawn(bob::run_until(bob_swap, is_xmr_locked));
